+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Make clock hands move smoothly.
 - Add dark theme.
+- Add clock scales.
 
 ### Changed
 - Change the clock style.